+++ pyomo/contrib/logical_expression_system/tests/test_logical_expression_system.py
@@ -1,6 +1,6 @@
 from itertools import product
 from copy import deepcopy
-from expression_system import \
+from pyomo.contrib.logical_expression_system import \
     (NotNode, LeafNode, OrNode, AndNode, IfNode, EquivalenceNode,
      bring_to_conjunctive_normal_form, is_conjunctive_normal_form,
      is_leaf_not_node, isNotNode, isOrNode)
